Add files via upload
This python script and macro will keep track of what tools that are in your tool changer. You will have to change your post so that your NC file shows the tool number and description. Also the script will look for "Vendor" as what machine it is. Soon as the script sees those it will open which ever excel file matches the Machine . Then the macro within the excel file will populate with the tools being used in the columns that match the tool number. Once the script is ran. It will monitor that NC file . Everytime you post a new NC file it will read it and open the correct excel file that matches to the machine. 
You will have to install Python , Watchdog ,PIP, and Pywin32 for the script to work correctly. Once the files have been downloaded place them in a folder on your desktop and open the script and the macro to change machine names, folder locations and so on and so fourth. Hope this helps.